--- src/main/java/com/girigiri/kwrental/equipment/Category.java
@@ -1,5 +1,0 @@
-package com.girigiri.kwrental.equipment;
-
-public enum Category {
-    CAMERA
-}
