+++ src/main/java/com/girigiri/kwrental/equipment/EquipmentRepository.java
@@ -1,5 +1,6 @@
 package com.girigiri.kwrental.equipment;
 
+import com.girigiri.kwrental.equipment.domain.Equipment;
 import java.util.Optional;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
