+++ src/main/java/com/girigiri/kwrental/equipment/domain/Category.java
@@ -1,0 +1,5 @@
+package com.girigiri.kwrental.equipment.domain;
+
+public enum Category {
+    CAMERA
+}
